Extract `getConfig`
