+++ src/pages/Home/index.js
@@ -1,10 +1,6 @@
 const HomePage = () => (
   <div>
-    <div>Agenda with pages</div>
-
-    <div>Go to /login</div>
-
-    <div>Go to /files/upload to test the file upload flow</div>
+    <p>Placeholder for the home page.</p>
   </div>
 );
